api/requirements.txt
praw==7.5.0
python-dotenv==0.19.0 